Movemos os metodos de JornadaAPI para JornadaController. Arrumamos a funcao que informa se o ponto esta dentro de um turno (nao cobria todos os casos antes)
